+++ src/Command/Redis/XREAD.php
@@ -46,6 +46,10 @@ class XREAD extends RedisCommand
 
     public function parseResponse($data)
     {
+        if (!$data) {
+            return [];
+        }
+
         $processedData = [];
 
         foreach ($data as $stream) {