+++ tests/Predis/Command/Redis/HGETDEL_Test.php
@@ -54,8 +54,9 @@ class HGETDEL_Test extends PredisCommandTestCase
 
     /**
      * @group connected
+     * @group relay-incompatible
      * @return void
-     * @requiresRedisVersion >= 7.9.0
+     * @requiresRedisVersion >= 8.0.0
      */
     public function testReturnsAndRemoveFieldsFromHash(): void
     {