+++ CHANGELOG.md
@@ -17,6 +17,11 @@
 - fix(stream): Fixed throwing `CommunicationException` when stream is EOF (#1548)
 - Removed automatic `conn_uid` parameter assignment (#1552)
 - fix(commands): Fixed wrong command API call on prefix processing (#1554)
+- fix `XREAD` response parsing while read null (#1563)
+- fix `XINFO` command responses parsing (#1560)
+
+### Maintenance
+- Updated Redis 8.0 test image (#1562)
 
 ### Maintenance
 - Added testing with 8.2 (#1567)